Added Ignore Rule for downloads/ Directory to .gitignore
Similar to blacklisted_tags.txt and tag_sets.txt, the downloads/ directory should not be included in the source code repository for the tool and thus it and its contents should be ignored. Added an ignore rule to .gitignore to accomplish this and prevent test data from being included in the repository.